9180 Move formats into companion objects
