+++ app/uk/gov/hmrc/tai/controllers/EmploymentsController.scala
@@ -22,7 +22,7 @@ import play.api.mvc.{Action, AnyContent, ControllerComponents}
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
 import uk.gov.hmrc.tai.controllers.predicates.AuthenticationPredicate
-import uk.gov.hmrc.tai.model.api.{ApiFormats, ApiResponse, EmploymentCollection}
+import uk.gov.hmrc.tai.model.api.{ApiResponse, EmploymentCollection}
 import uk.gov.hmrc.tai.model.domain.{AddEmployment, EndEmployment, IncorrectEmployment}
 import uk.gov.hmrc.tai.model.tai.TaxYear
 import uk.gov.hmrc.tai.service.EmploymentService
@@ -35,7 +35,7 @@ class EmploymentsController @Inject() (
   authentication: AuthenticationPredicate,
   cc: ControllerComponents
 )(implicit ec: ExecutionContext)
-    extends BackendController(cc) with ApiFormats with ControllerErrorHandler {
+    extends BackendController(cc) with ControllerErrorHandler {
 
   def employments(nino: Nino, year: TaxYear): Action[AnyContent] = authentication.async { implicit request =>
     employmentService
