+++ app/uk/gov/hmrc/tai/controllers/FileUploadController.scala
@@ -21,14 +21,13 @@ import play.api.libs.json.JsValue
 import play.api.mvc.{Action, ControllerComponents}
 import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
 import uk.gov.hmrc.tai.model.FileUploadCallback
-import uk.gov.hmrc.tai.model.api.ApiFormats
 import uk.gov.hmrc.tai.service.FileUploadService
 
 import scala.concurrent.Future
 
 @Singleton
 class FileUploadController @Inject() (fileUploadService: FileUploadService, cc: ControllerComponents)
-    extends BackendController(cc) with ApiFormats {
+    extends BackendController(cc) {
 
   def fileUploadCallback(): Action[JsValue] = Action.async(parse.json) { implicit request =>
     withJsonBody[FileUploadCallback] { fileUploadCallback =>
