+++ app/uk/gov/hmrc/tai/controllers/PersonController.scala
@@ -22,7 +22,7 @@ import play.api.mvc.{Action, AnyContent, ControllerComponents}
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
 import uk.gov.hmrc.tai.controllers.predicates.AuthenticationPredicate
-import uk.gov.hmrc.tai.model.api.{ApiFormats, ApiResponse}
+import uk.gov.hmrc.tai.model.api.ApiResponse
 import uk.gov.hmrc.tai.service.PersonService
 
 import scala.concurrent.ExecutionContext
@@ -34,7 +34,7 @@ class PersonController @Inject() (
   cc: ControllerComponents
 )(implicit
   ec: ExecutionContext
-) extends BackendController(cc) with ApiFormats {
+) extends BackendController(cc) {
 
   def person(nino: Nino): Action[AnyContent] = authentication.async { implicit request =>
     personService.person(nino) map { person =>
