+++ app/uk/gov/hmrc/tai/controllers/TotalTaxController.scala
@@ -22,7 +22,7 @@ import play.api.mvc.{Action, AnyContent, ControllerComponents}
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
 import uk.gov.hmrc.tai.controllers.predicates.AuthenticationPredicate
-import uk.gov.hmrc.tai.model.api.{ApiFormats, ApiResponse}
+import uk.gov.hmrc.tai.model.api.ApiResponse
 import uk.gov.hmrc.tai.model.tai.TaxYear
 import uk.gov.hmrc.tai.service.TotalTaxService
 import uk.gov.hmrc.tai.util.NpsExceptions
@@ -35,7 +35,7 @@ class TotalTaxController @Inject() (
   authentication: AuthenticationPredicate,
   cc: ControllerComponents
 )(implicit ec: ExecutionContext)
-    extends BackendController(cc) with ApiFormats with NpsExceptions with ControllerErrorHandler {
+    extends BackendController(cc) with NpsExceptions with ControllerErrorHandler {
 
   def totalTax(nino: Nino, year: TaxYear): Action[AnyContent] = authentication.async { implicit request =>
     totalTaxService.totalTax(nino, year) map { totalTax =>
