+++ app/uk/gov/hmrc/tai/controllers/income/IncomeController.scala
@@ -23,7 +23,7 @@ import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
 import uk.gov.hmrc.tai.controllers.ControllerErrorHandler
 import uk.gov.hmrc.tai.controllers.predicates.AuthenticationPredicate
-import uk.gov.hmrc.tai.model.api.{ApiFormats, ApiLink, ApiResponse}
+import uk.gov.hmrc.tai.model.api.{ApiLink, ApiResponse}
 import uk.gov.hmrc.tai.model.domain.TaxCodeIncomeComponentType
 import uk.gov.hmrc.tai.model.domain.income.TaxCodeIncomeStatus
 import uk.gov.hmrc.tai.model.domain.requests.UpdateTaxCodeIncomeRequest
@@ -40,7 +40,7 @@ class IncomeController @Inject() (
   cc: ControllerComponents
 )(implicit
   ec: ExecutionContext
-) extends BackendController(cc) with ApiFormats with ControllerErrorHandler {
+) extends BackendController(cc) with ControllerErrorHandler {
 
   def untaxedInterest(nino: Nino): Action[AnyContent] = authentication.async { implicit request =>
     incomeService.untaxedInterest(nino).map {
