+++ app/uk/gov/hmrc/tai/model/domain/Payment.scala
@@ -16,7 +16,7 @@
 
 package uk.gov.hmrc.tai.model.domain
 
-import play.api.libs.json.{Format, Json, OFormat}
+import play.api.libs.json.{Format, Json}
 import uk.gov.hmrc.tai.model.domain.formatters.EmploymentMongoFormatters
 
 import java.time.LocalDate
@@ -38,7 +38,7 @@ case class Payment(
 
 object Payment {
   implicit val paymentFrequencyFormat: Format[PaymentFrequency] = EmploymentMongoFormatters.paymentFrequencyFormat
-  implicit val format: OFormat[Payment] = Json.format[Payment]
+  implicit val format: Format[Payment] = Json.format[Payment]
 
   private implicit val LocalDateOrder: Ordering[LocalDate] = (x: LocalDate, y: LocalDate) => x.compareTo(y)
   implicit val dateOrdering: Ordering[Payment] = Ordering.by(_.date)
