+++ app/uk/gov/hmrc/tai/model/domain/PaymentFrequency.scala
@@ -16,8 +16,28 @@
 
 package uk.gov.hmrc.tai.model.domain
 
+import play.api.libs.json.{Format, JsResult, JsString, JsSuccess, JsValue}
+
 sealed trait PaymentFrequency {
   val value: String
+}
+
+object PaymentFrequency {
+  implicit val paymentFrequencyFormat: Format[PaymentFrequency] = new Format[PaymentFrequency] {
+    override def reads(json: JsValue): JsResult[PaymentFrequency] = json.as[String] match {
+      case "Weekly"      => JsSuccess(Weekly)
+      case "FortNightly" => JsSuccess(FortNightly)
+      case "FourWeekly"  => JsSuccess(FourWeekly)
+      case "Monthly"     => JsSuccess(Monthly)
+      case "Quarterly"   => JsSuccess(Quarterly)
+      case "BiAnnually"  => JsSuccess(BiAnnually)
+      case "Annually"    => JsSuccess(Annually)
+      case "OneOff"      => JsSuccess(OneOff)
+      case "Irregular"   => JsSuccess(Irregular)
+      case _             => throw new IllegalArgumentException("Invalid payment frequency")
+    }
+    override def writes(paymentFrequency: PaymentFrequency): JsValue = JsString(paymentFrequency.toString)
+  }
 }
 
 case object Weekly extends PaymentFrequency {
