--- app/uk/gov/hmrc/tai/model/tai/package.scala
@@ -1,29 +1,0 @@
-/*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.tai.model
-
-import play.api.libs.json._
-
-package object tai {
-  implicit val formatTaxYear: Format[TaxYear] = new Format[TaxYear] {
-    override def reads(j: JsValue): JsResult[TaxYear] = j match {
-      case JsNumber(n) => JsSuccess(TaxYear(n.toInt))
-      case x           => JsError(s"Expected JsNumber, found $x")
-    }
-    override def writes(v: TaxYear): JsValue = JsNumber(v.year)
-  }
-}
