--- test/resources/uk/gov/hmrc/tai/model/tai/TaiPackageSpec.scala
@@ -1,40 +1,0 @@
-/*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.tai.model.tai
-
-import org.scalatestplus.play.PlaySpec
-import play.api.libs.json.{JsError, JsNumber, JsString, JsSuccess}
-
-class TaiPackageSpec extends PlaySpec {
-  "TaiPackage" should {
-    "deserialise a Tax Year" when {
-      "given the correct Json type" in {
-        formatTaxYear.reads(JsNumber(2017)) mustBe JsSuccess(TaxYear(2017))
-      }
-    }
-
-    "return as JsError" when {
-      "given the wrong Json type" in {
-        formatTaxYear.reads(JsString("2017")) mustBe JsError("Expected JsNumber, found \"2017\"")
-      }
-    }
-
-    "serialise a TaxYear instance" in {
-      formatTaxYear.writes(TaxYear(2017)) mustBe JsNumber(2017)
-    }
-  }
-}
